AI/digital_sig/requirements.txt
PyMuPDF
fastapi
opencv-python
numpy
pycryptodome
